docs(compiler): clarify CLI usage and clean tooling
Update the README to show the actual rustcc binary contract and keep --chapter scoped to the official test harness. Add a small example C file so the documented commands are executable, then make current clippy-all warnings clean without touching the official tests.\n\nConstraint: W22-T1 is documentation/tooling polish only; full W22 regression remains a separate task.\nConstraint: tests/ harness contents must remain unchanged.\nRejected: Document examples/hello.c without adding the file | adversarial gate proved the command failed.\nConfidence: high\nScope-risk: narrow\nDirective: Do not claim W22-T2 full regression from this commit; it only proves build/test/clippy and documented invocation examples.\nTested: git diff -- tests empty; cargo fmt --all -- --check; cargo build --release no warnings; cargo test --release (10 passed); cargo clippy --release -- -W clippy::all warning-clean; README example commands for examples/hello.c; final code review APPROVE; final adversarial gate APPROVE.\nNot-tested: W22-T2 full regression matrix; Final F1-F4 review wave.

+++ src/parse/parser.rs
@@ -690,7 +690,7 @@ impl Parser {
 
     fn expect_array_size(&mut self) -> Result<usize> {
         let raw = match &self.peek().kind {
-            TokenKind::Constant(value) => i64::from(*value),
+            TokenKind::Constant(value) => *value,
             TokenKind::LongConstant(value) | TokenKind::UIntConstant(value, _) => *value,
             TokenKind::CharLiteral(value) => i64::from(*value),
             _ => bail!("parse error: expected constant array size"),
@@ -1077,7 +1077,7 @@ impl Parser {
     fn parse_unary_expr(&mut self) -> Result<Expr> {
         let expr = match &self.peek().kind {
             TokenKind::Constant(value) => {
-                let value = i64::from(*value);
+                let value = *value;
                 self.current += 1;
                 Expr::Constant(value)
             }
